requirements.txt
accelerate==0.31.0
annotated-types==0.7.0
anyio==4.4.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1717618050233/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
diffusers==0.29.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
ExifRead==3.0.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock @ file:///croot/filelock_1700591183607/work
fsspec==2024.6.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==7.1.0
Jinja2 @ file:///croot/jinja2_1716993405101/work
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1717597493534/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
ollama==0.2.1
orjson==3.10.5
packaging==24.1
peft==0.11.1
pillow @ file:///croot/pillow_1714398848491/work
psutil==5.9.8
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML @ file:///croot/pyyaml_1698096049011/work
regex==2024.5.15
requests @ file:///croot/requests_1716902831423/work
rich==13.7.1
safetensors==0.4.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
sympy @ file:///croot/sympy_1701397643339/work
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typer==0.12.3
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
ujson==5.10.0
urllib3 @ file:///croot/urllib3_1715635851070/work
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
zipp==3.19.2
